llm/requirements.txt
langchain
openai
python-dotenv
beautifulsoup4
unstructured
chromadb
tiktoken